Split the humongous types.go into smaller and more focused files
Also, get rid of code duplication in unmarshalling keys which take one or multiple values
